moves camera monitor speaking to its own component, adds voice transformation, and adds speech to televisions
