+++ Content.Server/SurveillanceCamera/Systems/SurveillanceCameraMicrophoneSystem.cs
@@ -25,8 +25,7 @@ public sealed class SurveillanceCameraMicrophoneSystem : EntitySystem
             return;
         }
 
-        var wrapped = Loc.GetString("surveillance-camera-microphone-message", ("speaker", speaker), ("message", message));
-        var ev = new SurveillanceCameraSpeechSendEvent(speaker, message, wrapped);
+        var ev = new SurveillanceCameraSpeechSendEvent(speaker, message);
 
         foreach (var monitor in camera.ActiveMonitors)
         {
@@ -38,14 +37,12 @@ public sealed class SurveillanceCameraMicrophoneSystem : EntitySystem
 public sealed class SurveillanceCameraSpeechSendEvent : EntityEventArgs
 {
     public EntityUid Speaker { get; }
-    public string OriginalMessage { get; }
-    public string WrappedMessage { get; }
+    public string Message { get; }
 
-    public SurveillanceCameraSpeechSendEvent(EntityUid speaker, string originalMessage, string wrappedMessage)
+    public SurveillanceCameraSpeechSendEvent(EntityUid speaker, string message)
     {
         Speaker = speaker;
-        OriginalMessage = originalMessage;
-        WrappedMessage = wrappedMessage;
+        Message = message;
     }
 }
