+++ Resources/Locale/en-US/surveillance-camera/surveillance-camera-speaker.ftl
@@ -1,0 +1,1 @@
+surveillance-camera-microphone-message = {$speaker} [{$originalName}]
