+++ Resources/Locale/en-US/surveillance-camera/surveillance-camera-ui.ftl
@@ -11,4 +11,3 @@ surveillance-camera-monitor-ui-no-subnets = No Subnets
 surveillance-camera-setup = Setup
 surveillance-camera-setup-ui-set = Set
 
-surveillance-camera-microphone-message = {$speaker} says, "{$message}"
