+++ Resources/Prototypes/Entities/Structures/Machines/Computers/computers.yml
@@ -675,6 +675,7 @@
       range: 200
     - type: DeviceNetworkRequiresPower
     - type: Speech
+    - type: SurveillanceCameraSpeaker
     - type: SurveillanceCameraMonitor
       speechEnabled: true
     - type: ActivatableUI
